Movie-Recommendation-System-main/requirements.txt
streamlit
numpy
Pillow
pandas
requests
beautifulsoup4
urllib3
